fix: sign manifest with cosign (#245)
--enforce-container-sigpolicy is failing due to a missing signature.

This PR signs the manifests as well as the image, which should allow us
to enforce signatures during rebases and updates.

Due to running the manifest step in a container, and Cosign not working
well in said container, I've moved the sign step to a separate job which
is completed after pushing the manifest.

---------

Co-authored-by: Tulip Blossom <tulilirockz@outlook.com>