feat: add nvtop and use epel version of staging copr for packages (#319)
This makes it so we can use the EPEL version instead of raw
centos-stream for building our packages, makes it so we can build A LOT
more stuff on our copr instead of being just stuff restricted to the
centos repos. Itll also allow us to just disable the centos-stream
chroot too, so thats pretty nice!