feat(gdx): use negativo for nvidia driver builds + enable arm builds (#316)
This allows for x64 AND aarch64 builds for the nvidia drivers! It also
uses akmods so its a bit cleaner to build and we could eventually
upstream this effort to a `centos-hwe` thing if we really need to.

This seems to be blocked because of missing things upstream. Mainly
typos, so it should be relatively quick to fix! Check out the actions