Fix image tag discovery for changelogs and changelog title format (#639)
This PR fixes the tag discovery for generating the changelogs.

Due to a change in the tag format from `{target}-{numbers}` to
`{target}.{numbers}` the regex needed to be updated in order to pull in
new versions.

While I was there, and from conversations in Discord, I adjusted the
formatting of the version in the title to be `c{version}s.{date}` from
the old `c{version}.{date}.s`.

Tested locally with `python changelog.py lts out.txt changelog.md`